Turn heights from ints to floats

+++ Honours Project/Assets/Architect/BlockProperties.cs
@@ -4,13 +4,12 @@ using UnityEngine;
 
 namespace Architect
 {
-	//[CreateAssetMenu(fileName = "NewBlockProperties", menuName = "Architect/Block Properties", order = 2)]
 	[System.Serializable]
 	public class BlockProperties
 	{
 		//public Sprite sprite; // TEMP
 		public int width;
-		public int height;
+		public float height;
 		public List<Style> styles = new List<Style>();
 	}
